feat(new_bin_bench): Merge lib_bench::Assistant and bin_bench::Assistant ...
* Cleanup old code
* Make use of ModulePath instead of using a String
