+++ iai-callgrind-runner/src/api.rs
@@ -227,6 +227,7 @@ pub struct LibraryBenchmarkConfig {
 
 #[derive(Debug, Default, Serialize, Deserialize)]
 pub struct LibraryBenchmarkGroup {
+    // TODO: SHOULD BE A STRING and not an Option
     pub id: Option<String>,
     pub config: Option<LibraryBenchmarkConfig>,
     pub compare: bool,
